pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-login>=0.6.3",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "google-genai>=1.32.0",
    "stripe>=12.5.0",
    "requests>=2.32.5",
    "python-telegram-bot>=22.3",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
